requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
influxdb==5.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
msgpack==0.6.1
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.12.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
